logger: move to pkg (#2031)
